Added main package and Options
main package is now the home of AppMain, ArgumentLineHandler, and two
new classes, Options and OptionsIO.

Options stores global variables used by the program.
OptionsIO handles creating/reading from an ini file located in the main
directory of the project.  If an "options.ini" file is not found, a new
one will be created with default values.  For now, any new options will
have to be added into a couple places: Options class,
writeDefaultsToNewFile(), and readFile().  I'll come up with a better
way of doing it in the future.